Core: make Repl an interface
The previous refactoring still did not allow the flexibility required for
different kinds of clients. Therefore, Repl has now been refactored into an
interface with the only method all implementations have in common: eval. Reading
input is now completely up to the client and not bound to any contract in any
way.

This also means that there cannot be an exit command in the core module anymore,
as every REPL is now substantially different than any other.