Eclipse: properly schedule hooks on the UI thread
As noted in the Javadoc for the hooks: since hooks are called automatically by
the commands, which in turn are called by the REPL running in its own thread,
hooks try to update the UI from another thread than the UI thread. The new UIJob
implementation actually under the hood calls asyncExe like we did before, but it
has a bit more monitoring going on as well.