Core: make commands return their hook
As mentiond in the IHook interface, having the commands return their hooks
allows for clients to process them whenever (e.g. after a certain event, or
immediately, et cetera) and wherever (e.g. on the UI thread in Eclipse) they
need.

It also resolves the issue that Guice needs to instantiate the hooks when it
instantiates the commands, whilst the EclipseDisplay (required for the hooks to
be instantiated) cannot yet be created due to the Composite argument not being
available yet.